+++ src/api/java/appeng/api/definitions/IBlocks.java
@@ -148,6 +148,8 @@ public interface IBlocks
 
 	ITileDefinition iface();
 
+	ITileDefinition exposer();
+
 	ITileDefinition cellWorkbench();
 
 	ITileDefinition iOPort();